feat(auth): complete user authentication – final commit

+++ internal/api/workout_handler.go
@@ -3,9 +3,11 @@ package api
 import (
 	"database/sql"
 	"encoding/json"
+	"errors"
 	"log"
 	"net/http"
 
+	"github.com/aliSHAIKH001/api_go/internal/middleware"
 	"github.com/aliSHAIKH001/api_go/internal/store"
 	"github.com/aliSHAIKH001/api_go/internal/utils"
 )
@@ -21,13 +23,13 @@ type WorkoutHandler struct {
 func NewWorkoutHandler(workoutStore store.WorkoutStore, logger *log.Logger) *WorkoutHandler {
 	return &WorkoutHandler{
 		workoutStore: workoutStore,
-		logger: logger,
+		logger:       logger,
 	}
 }
 
 func (wh *WorkoutHandler) HandleGetWorkoutById(w http.ResponseWriter, r *http.Request) {
 
-	// ReadIDParam gets the id from the request and parses it into int64 
+	// ReadIDParam gets the id from the request and parses it into int64
 	workoutID, err := utils.ReadIDParam(r)
 	if err != nil {
 		wh.logger.Printf("ERROR: readIDParam: %v", err)
@@ -41,7 +43,7 @@ func (wh *WorkoutHandler) HandleGetWorkoutById(w http.ResponseWriter, r *http.Re
 		utils.WriteJSON(w, http.StatusInternalServerError, utils.Envelope{"error": "internal server error"})
 		return
 	}
-	
+
 	// WriteJSON sets the headers and status codes and pretty prints JSON.
 	utils.WriteJSON(w, http.StatusOK, utils.Envelope{"workout": workout})
 }
@@ -54,7 +56,15 @@ func (wh *WorkoutHandler) HandleCreateWorkout(w http.ResponseWriter, r *http.Req
 		utils.WriteJSON(w, http.StatusBadRequest, utils.Envelope{"error": "invalid request sent"})
 		return
 	}
-	
+
+	// To get the ID of the user who is creating the workout
+	currentUser := middleware.GetUser(r)
+	if currentUser == nil || currentUser == store.AnonymousUser {
+		utils.WriteJSON(w, http.StatusBadRequest, utils.Envelope{"error": "you must be logged in"})
+		return
+	}
+	workout.UserID = currentUser.ID
+
 	// Created workout is different from the workout as it has an ID field returned by the query in CreateWorkout func
 	createdWorkout, err := wh.workoutStore.CreateWorkout(&workout)
 	if err != nil {
@@ -67,18 +77,21 @@ func (wh *WorkoutHandler) HandleCreateWorkout(w http.ResponseWriter, r *http.Req
 }
 
 func (wh *WorkoutHandler) HandleUpdateWorkoutByID(w http.ResponseWriter, r *http.Request) {
+
+	// ReadIDParam gets the id from the request and parses it into int64
 	workoutID, err := utils.ReadIDParam(r)
 	if err != nil {
 		wh.logger.Printf("ERROR: readIDParam: %v", err)
 		utils.WriteJSON(w, http.StatusBadRequest, utils.Envelope{"error": "invalid workout update ID"})
-		return 
+		return
 	}
 
+	// Getting the existing workout to update it
 	existingWorkout, err := wh.workoutStore.GetWorkoutByID(workoutID)
 	if err != nil {
 		wh.logger.Printf("ERROR: getWorkoutByID: %v", err)
 		utils.WriteJSON(w, http.StatusInternalServerError, utils.Envelope{"error": "internal server error"})
-		return 
+		return
 	}
 
 	if existingWorkout == nil {
@@ -88,22 +101,24 @@ func (wh *WorkoutHandler) HandleUpdateWorkoutByID(w http.ResponseWriter, r *http
 
 	// The zero value for pointers is nil making it easier to check if the client actually sent something.
 	var updateWorkoutRequest struct {
-		Title *string `json:"title"`
-		Description *string `json:"description"`
-		DurationMinutes *int `json:"duration_minutes"`
-		CaloriesBurned *int `json:"calories_burned"`
-		Entries []store.WorkoutEntry `json:"entries"`
+		Title           *string              `json:"title"`
+		Description     *string              `json:"description"`
+		DurationMinutes *int                 `json:"duration_minutes"`
+		CaloriesBurned  *int                 `json:"calories_burned"`
+		Entries         []store.WorkoutEntry `json:"entries"`
 	}
 
+	// Decode the request body into the updateWorkoutRequest struct
 	err = json.NewDecoder(r.Body).Decode(&updateWorkoutRequest)
 	if err != nil {
 		wh.logger.Printf("ERROR: decodingUpdateRequest: %v", err)
 		utils.WriteJSON(w, http.StatusBadRequest, utils.Envelope{"error": "invalid request payload"})
 		return
 	}
+
 	// MOdifies the Exising workuot for only the fields given in the request
 	if updateWorkoutRequest.Title != nil {
-		existingWorkout.Title = *updateWorkoutRequest.Title	
+		existingWorkout.Title = *updateWorkoutRequest.Title
 	}
 	if updateWorkoutRequest.Description != nil {
 		existingWorkout.Description = *updateWorkoutRequest.Description
@@ -118,6 +133,28 @@ func (wh *WorkoutHandler) HandleUpdateWorkoutByID(w http.ResponseWriter, r *http
 		existingWorkout.Entries = updateWorkoutRequest.Entries
 	}
 
+	// Prevents unauthorized users from updating the workout
+	currentUser := middleware.GetUser(r)
+	if currentUser == nil || currentUser == store.AnonymousUser {
+		utils.WriteJSON(w, http.StatusBadRequest, utils.Envelope{"error": "you must be logged in"})
+		return
+	}
+	WorkoutOwner, err := wh.workoutStore.GetWorkoutOwner(workoutID)
+	if err != nil {
+		if errors.Is(err, sql.ErrNoRows) {
+			utils.WriteJSON(w, http.StatusNotFound, utils.Envelope{"error": "workout does not exist"})
+			return
+		}
+
+		utils.WriteJSON(w, http.StatusInternalServerError, utils.Envelope{"error": "internal server error"})
+		return
+	}
+	if WorkoutOwner != currentUser.ID {
+		utils.WriteJSON(w, http.StatusForbidden, utils.Envelope{"error": "you are not authorized to update this workout"})
+		return
+	}
+
+	// Finally we update the existing workout in the database.
 	err = wh.workoutStore.UpdateWorkout(existingWorkout)
 	if err != nil {
 		wh.logger.Printf("ERROR: updatingWorkout: %v", err)
@@ -134,17 +171,39 @@ func (wh *WorkoutHandler) HandleDeleteWorkoutByID(w http.ResponseWriter, r *http
 	if err != nil {
 		wh.logger.Printf("ERROR: readIDParam: %v", err)
 		utils.WriteJSON(w, http.StatusBadRequest, utils.Envelope{"error": "invalid workout delete ID"})
-		return 
+		return
 	}
 
-	err	= wh.workoutStore.DeleteWorkout(workoutID)
+	// Prevents unauthorized users from deleting the workout
+	currentUser := middleware.GetUser(r)
+	if currentUser == nil || currentUser == store.AnonymousUser {
+		utils.WriteJSON(w, http.StatusBadRequest, utils.Envelope{"error": "you must be logged in"})
+		return
+	}
+	WorkoutOwner, err := wh.workoutStore.GetWorkoutOwner(workoutID)
+	if err != nil {
+		if errors.Is(err, sql.ErrNoRows) {
+			utils.WriteJSON(w, http.StatusNotFound, utils.Envelope{"error": "workout does not exist"})
+			return
+		}
+
+		utils.WriteJSON(w, http.StatusInternalServerError, utils.Envelope{"error": "internal server error"})
+		return
+	}
+	if WorkoutOwner != currentUser.ID {
+		utils.WriteJSON(w, http.StatusForbidden, utils.Envelope{"error": "you are not authorized to delete this workout"})
+		return
+	}
+
+	// Delete the workout from the database
+	err = wh.workoutStore.DeleteWorkout(workoutID)
 	if err == sql.ErrNoRows {
 		http.Error(w, "workout not found", http.StatusNotFound)
 		return
 	}
 	if err != nil {
 		http.Error(w, "Error deleting workout", http.StatusInternalServerError)
-		return 
+		return
 	}
 
 	w.WriteHeader(http.StatusNoContent)
